configure: Remove --with-group option for device locking
This option is not used anywhere in the code or the build system, except
that ./configure will not build device locking support if this is set to
a non-existent group. (It chooses uucp by default.)

However, that check does not actually work. It attempts to run the chgrp
command on a temporary file; but since ./configure is run without root
privileges, that always fails unless the user running ./configure is a
member of the group.

Remove both the option and the broken check. It is assumed that device
locking support was typically disabled in builds because of this issue,
and it will now be restored except where --disable-locking is specified.