CI: Do not use silent rules
Instead use make(1) env MAKEFLAGS set to `-w -O`.

With silent rules we don't see compile commands and only see a error
messages. This is beautiful when there is no errors but harder to
understand the cause when there are errors.

For example, after adding `-Wno-error=unused-but-set-variable`
I still see the error:

  CXX      genesys/libgenesys_la-device.lo
genesys/genesys.cpp:843:19: error: variable 'i' set but not used [-Werror,-Wunused-but-set-variable]
  843 |     for (unsigned i = 0;; ++i) {
      |                   ^

But why? Hard to tell since we don't see compile command line.

With `-w -O` we see each compilation job grouped in the log.

This will help to debug build issues in CI, where reproduction is
costly/harder for a casual contributor. Compilation command is just
above the error. Find errors in log by searching for `error:` or
`warning:` strings.