saned: add option to allow relaying network scanners
My intended use case is the following:
I want to compile saned backends with different libraries (for example
glibc) than a specific application (as if it were a chroot of another
distro version). This application cannot dlopen sane
backends because of the library mismatch.
The specific application will be
built with a stripped down libsane that only has the net backend,
and therefore very few dependencies. The net backend points to saned.
saned has a full config and set of backends, and is responsible for
exporting all scanners to the application. This option allows to
reexport network scanners this way. I only have to make sure saned does
not contact itself.