backend/kodakio.c: Remove unnecessary Unicode character
There is an bidirectional Unicode character <202d> in the comment.
The character looks like a typo, because it does not look fitting
there, so IMHO it can be removed.

Checking tools like rpminspect are wary of Unicode characters in the
code after CVE regarding bidirectional Unicode characters from several
years ago, and character <202d> is for left-to-right override in
Unicode, so tools report them as an issue.