saned: use pidfd to avoid unnecessary frequent wake-ups
Avoid waking up every 500ms in standalone mode to reap any terminated
child processes if pidfds are available. This benefits Linux users with
kernel version 5.3 and glibc 2.36 or later.

Introduces a linked list of fds combining listening sockets and pidfds
and creates a poll set from this list as required.

Child process tracking is now generalised so that the Avahi service is
handled in the same way as client connection processes.