avision: add support for Xerox Book Scanner 4167
The Xerox Book Scanner 4167 is a rebadged Avision FB6080E with modified USB IDs. It works perfectly with the existing support in the avision backend for the Avision FB6080E.
There is very little information about this model available online. From what I could find, it seems this model was distributed as a component, along with a pre-configured PC running Scannx software, of the Xerox BookCentre S7141 turn-key institutional book scanning system.